% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSearchTerms.R
\name{abbrevSciName}
\alias{abbrevSciName}
\title{Abbreviates scientific names}
\usage{
abbrevSciName(names)
}
\arguments{
\item{sci_names}{a vector of scientific names to be abbreviated}

\item{ids}{a vector the same length as \code{sci_names} containing the
internalTaxonId for each sci_name}
}
\value{
a dataframe with all the abbreviations for each sci name matched
to the correct id
}
\description{
Abbreviates scientific names
}
