% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikiPageViews.R
\name{checkAccess}
\alias{checkAccess}
\title{check access method is acceptable}
\usage{
checkAccess(access)
}
\arguments{
\item{access}{the access method}
}
\description{
check access method is acceptable
}
