% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikiPageViews.R
\name{checkDate}
\alias{checkDate}
\title{check dates}
\usage{
checkDate(year, month, day, start_or_end)
}
\arguments{
\item{year}{year}

\item{month}{month}

\item{day}{day}

\item{start_or_end}{whether it's the start date or the end date}
}
\description{
check whether the given dates are real
}
