% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WoSSearch.R
\name{checkExportMessage}
\alias{checkExportMessage}
\title{Checks for WoS no results error message and closes if present}
\usage{
checkExportMessage(remDr)
}
\arguments{
\item{remDr}{remote driver for rselenium}
}
\description{
Checks for WoS no results error message and closes if present
}
\section{Used in}{
 \code{\link{WoSExport}}
}

