% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WoSSearch.R
\name{closeRemDr}
\alias{closeRemDr}
\title{closes remDr and rD}
\usage{
closeRemDr(remDr, rD)
}
\arguments{
\item{remDr}{remote driver for rselenium}

\item{rD}{also remote driver? for rselenium}
}
\description{
closes remDr and rD
}
\section{Used in}{
 \code{\link{WoSSearch}}
}

