% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleSearch.R
\name{createGoogleUrls}
\alias{createGoogleUrls}
\title{Make google urls from names for api queries}
\usage{
createGoogleUrls(query, api, custom_search, site_restricted = FALSE)
}
\description{
Make google urls from names for api queries
}
