% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSearchTerms.R
\name{createSearchTerms}
\alias{createSearchTerms}
\title{Creates a Web of Science search term from common and scientific names}
\usage{
createSearchTerms(
  names,
  name_types_to_include = "all",
  remove_unnecessary_terms = T,
  remove_overlapping_names = T,
  include_sci_name_abbreviation = F
)
}
\arguments{
\item{names}{tibble of names you want to change into search terms, created
using one of the format functions}

\item{name_types_to_include}{vector of the name types you want to include in
the search terms}

\item{remove_overlapping_names}{logical, defaults to F, if true any names
that are not a main name and are contained within the names of another
species are removed}

\item{include_sci_name_abbreviation}{logical, defaults to F, if true the
abbreviation for each scientific name will be included}
}
\value{
a tibble with columns \code{ID} and one column for each site's search
term
}
\description{
Creates a Web of Science search term from common and scientific names
}
