% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eolUrlScrapeR.R
\name{EOLDeduplicate}
\alias{EOLDeduplicate}
\title{Remove duplicate EOL urls}
\usage{
EOLDeduplicate(species_data)
}
\arguments{
\item{EOL_urls}{tibble created by \code{\link{EOLUrlScrapeR}}}

\item{names}{tibble created by binding IOC names to names from tidyRedlist}
}
\value{
a tibble with columns \code{internalTaxonId}, \code{text} and
\code{link} with duplicates removed
}
\description{
\code{EOLDeduplicate} removes any cases where EOL urls have been assigned
to a species incorrectly because its name is contained within another
species name
}
\section{Requires}{
 \code{\link{rselenium}}
}

