% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eolNameScrapeR.R
\name{EOLSciNameScrapeR}
\alias{EOLSciNameScrapeR}
\title{Scrape scientific names from Encyclopedia of Life}
\usage{
EOLSciNameScrapeR(EOL_urls)
}
\arguments{
\item{EOL_urls}{tibble of EOL urls created by \code{\link{EOLUrlScrapeR}}}
}
\value{
a tibble with columns \code{internalTaxonId}, \code{scientificName},
\code{name}, \code{main}, \code{language} and \code{source}.
}
\description{
\code{EOLSciNamesScrapeR} takes urls for species homepages and scrapes
the alternate latin names for each spp
}
\section{Requires}{
 rvest
}

