% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eolUrlScrapeR.R
\name{EOLUrlScrapeR}
\alias{EOLUrlScrapeR}
\title{Finds Encyclopedia of Life urls for each species}
\usage{
EOLUrlScrapeR(names, best_match)
}
\arguments{
\item{names}{a tibble with all names associated with their internalTaxonIds}

\item{best_match}{logical, if true, will return the link for which the
text best matches the latin name searched, if false, will return all the
dropdown results for each search term}
}
\value{
a tibble with columns \code{internalTaxonId}, \code{text} and
\code{link}
}
\description{
\code{EOLUrlScrapeR} enters each value in search_terms into the eol search
bar and returns the links and text from the resultant dropdown menu it
returns a dataframe with link text and link for each search term, will
return NA if no curated results for search.
}
\section{Requires}{
 \code{\link{rselenium}}
}

