% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_fit_map.R
\name{estimate_candidate_gain_map}
\alias{estimate_candidate_gain_map}
\title{Find a map that satisfies the total efficiency gain relationship.}
\usage{
estimate_candidate_gain_map(
  efficiency,
  prior_type,
  params_detection,
  neural_resource,
  g_min,
  g_max,
  a,
  b,
  c,
  d
)
}
\arguments{
\item{cp}{}
}
\value{

}
\description{
Find a map that satisfies the total efficiency gain relationship.
}
