% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eolUrlScrapeR.R
\name{ExtractHyperlink}
\alias{ExtractHyperlink}
\title{Extracts text and url from hyperlink}
\usage{
ExtractHyperlink(x_path, remDr, id)
}
\arguments{
\item{x_path}{xpath for hyperlink}

\item{remDr}{rselenium remote driver}

\item{id}{the id associated with the term searched}
}
\value{
a tibble with columns \code{internalTaxonId}, \code{text} and
\code{link}
}
\description{
this function extracts the link text and url from an EOL dropdown list,
used in eolurlscraper it returns a dataframe with link text and link for
each entry in the EOL dropdown list
}
\section{Requires}{
 \code{\link{rselenium}}
}

