% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_fit_map.R
\name{f.growth}
\alias{f.growth}
\title{Weibull growth. Customized to have a front and back end.
See paper Walshe & Geisler (2021) for details}
\usage{
f.growth(g_min, g_max, a, b, c, d, eccentricity)
}
\arguments{
\item{c_2}{}
}
\value{

}
\description{
Weibull growth. Customized to have a front and back end.
See paper Walshe & Geisler (2021) for details
}
