% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_fit_map.R
\name{f.optim}
\alias{f.optim}
\title{The objective to ensure the gain map satisfies the resource constraint.}
\usage{
f.optim(g_min, g_max, a, b, c, d, efficiency, neural_resource, G)
}
\arguments{
\item{c}{}
}
\value{

}
\description{
The objective to ensure the gain map satisfies the resource constraint.
}
