% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WoSSearch.R
\name{findAndClick}
\alias{findAndClick}
\title{Find an element and click it in rselenium}
\usage{
findAndClick(remDr, xpath)
}
\arguments{
\item{remDr}{remote driver for rselenium}

\item{xpath}{xpath for element to be clicked}
}
\description{
finds element specified by \code{xpath} and clicks it
}
\section{Used in}{
 \code{\link{WoSExport}}, \code{\link{WoSInitialExportSetup}} and
\code{\link{WoSSubsequentExportSetup}}
}

