% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOCBirdList.R
\name{fixNamesWithBrackets}
\alias{fixNamesWithBrackets}
\title{splits names with brackets into two names}
\usage{
fixNamesWithBrackets(names)
}
\arguments{
\item{names}{a names dataframe}
}
\description{
if the name is A (b) c, will return the names A c and A b c, then add these
to the overall names list and remove duplicates
}
