% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_optimizer_results.R
\name{format_optimizer_results}
\alias{format_optimizer_results}
\title{Supply an environment to the function. Format it back to something acceptable.
Note. This is a very custom function. It accepts an environment that is output from a estimation job.
This is a brittle function and should be used with caution if any changes to the optimization
procedure are made.}
\usage{
format_optimizer_results(optimized_results)
}
\arguments{
\item{optimized_results}{}
}
\value{

}
\description{
Supply an environment to the function. Format it back to something acceptable.
Note. This is a very custom function. It accepts an environment that is output from a estimation job.
This is a brittle function and should be used with caution if any changes to the optimization
procedure are made.
}
