% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redlistNamesToInput.R
\name{formatID_name_and_taxon_id}
\alias{formatID_name_and_taxon_id}
\title{Formats ID_name_and_taxon_id into names tibble}
\usage{
formatID_name_and_taxon_id(ID_name_and_taxon_id)
}
\arguments{
\item{ID_name_and_taxon_id}{the name being used as an ID and IUCN internal
taxon ID}
}
\value{
a tibble with columns: \code{ID}, the unique identifier for the
species (in this case the main scientific name); \code{name}, the names of
the species; \code{name_type}, the type of name (in this case scientific or
common); \code{language}, the language of the name; and \code{main}, logical,
whether the name is to be used as the main name for the species
}
\description{
Formats ID_name_and_taxon_id into names tibble
}
