% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redlistNamesToInput.R
\name{formatIUCN}
\alias{formatIUCN}
\title{Formats names data from IUCN Red List download into format required for
creating search terms}
\usage{
formatIUCN(iucn_synonyms, iucn_common_names, iucn_simple_summary)
}
\arguments{
\item{iucn_synonyms}{synonyms csv downloaded from IUCN}

\item{iucn_common_names}{common names csv downloaded from IUCN}

\item{iucn_simple_summary}{simple summary csv downloaded from IUCN}
}
\value{
a tibble with columns: \code{ID}, the unique identifier for the
species (in this case the main scientific name); \code{name}, the names of
the species; \code{name_type}, the type of name (in this case scientific or
common); \code{language}, the language of the name; and \code{main}, logical,
whether the name is to be used as the main name for the species
}
\description{
Formats names data from IUCN Red List download into format required for
creating search terms
}
