% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redlistNamesToInput.R
\name{formatIUCNDirectory}
\alias{formatIUCNDirectory}
\title{Formats IUCN Red List download directory into format required for
creating search terms}
\usage{
formatIUCNDirectory(iucn_wd)
}
\arguments{
\item{iucn_wd}{the working directory for IUCN download data}
}
\value{
a tibble with columns: \code{ID}, the unique identifier for the
species (in this case the main scientific name); \code{name}, the names of
the species; \code{name_type}, the type of name (in this case scientific or
common); \code{language}, the language of the name; and \code{main}, logical,
whether the name is to be used as the main name for the species
}
\description{
Formats IUCN Red List download directory into format required for
creating search terms
}
