% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redlistNamesToInput.R
\name{formatIUCNSynonyms}
\alias{formatIUCNSynonyms}
\title{Formats synonyms data from IUCN Red List download into format required for
creating search terms}
\usage{
formatIUCNSynonyms(ID_name_and_taxon_id, iucn_synonyms)
}
\arguments{
\item{ID_name_and_taxon_id}{the name being used as an ID and IUCN internal
taxon ID}

\item{iucn_synonyms}{synonyms csv downloaded from IUCN}
}
\value{
a tibble with columns: \code{ID}, the unique identifier for the
species (in this case the main scientific name); \code{name}, the names of
the species; \code{name_type}, the type of name (in this case scientific or
common); \code{language}, the language of the name; and \code{main}, logical,
whether the name is to be used as the main name for the species
}
\description{
Formats synonyms data from IUCN Red List download into format required for
creating search terms
}
