% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikiScrapeR.R
\name{formatNamesToWiki}
\alias{formatNamesToWiki}
\title{make names into wikipedia url format}
\usage{
formatNamesToWiki(main_names)
}
\arguments{
\item{main_names}{a tibble with the ID of each species and its main name
(scientific is best to minimise ambiguity)}
}
\value{
the same tibble with column 'search' added containing formatted
names
}
\description{
make names into wikipedia url format
}
