% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_fit_map.R
\name{full_objective}
\alias{full_objective}
\title{Create the objective function used to solve the map.}
\usage{
full_objective(a, b, c, d, g_min, g_max, efficiency)
}
\arguments{
\item{efficiency}{}
}
\value{

}
\description{
Create the objective function used to solve the map.
}
