% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSeleniumHelperFunctions.R
\name{getElementText}
\alias{getElementText}
\title{Find an element and get text}
\usage{
getElementText(remDr, using, path)
}
\arguments{
\item{remDr}{the remote driver}

\item{using}{the type of path to use}

\item{path}{the path to use}
}
\value{
a value containing the text from the element
}
\description{
Find an element and get text
}
