% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOCBirdList.R
\name{getIOCNames}
\alias{getIOCNames}
\title{adds IOC names to existing names tibble}
\usage{
getIOCNames(species_data)
}
\arguments{
\item{species_data}{IUCN Redlist data cleaned with tidyRedlist}
}
\description{
adds IOC names to existing names tibble
}
