% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redlistNamesToInput.R
\name{getIUCNFiles}
\alias{getIUCNFiles}
\title{Get files from IUCN working directory}
\usage{
getIUCNFiles(iucn_wd)
}
\value{
a list with simple summary, common names and synonyms
}
\description{
Get files from IUCN working directory
}
