% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSearchTerms.R
\name{googleSearchTerms}
\alias{googleSearchTerms}
\title{Creates a Google search term from common and scientific names}
\usage{
googleSearchTerms(names_to_search)
}
\arguments{
\item{names}{tibble with at least cols \code{ID}, \code{name}}
}
\value{
a tibble with columns \code{internalTaxonId} and \code{google}
which contains google search terms
}
\description{
Creates a Google search term from common and scientific names
}
