% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WoSSearch.R
\name{importRefs}
\alias{importRefs}
\title{Import downloaded .bib references}
\usage{
importRefs(directory, search_ids, exported)
}
\arguments{
\item{directory}{the directory to be used for downloaded search results}

\item{search_ids}{a vector/ column the same length as \code{search_ids}
containing an ID to be used to identify search results to each search term}
}
\value{
the number of files that have been exported
}
\description{
imports all downloaded .bib files in \code{directory} and associates
them with the correct \code{search_ids} using \code{exports}
}
\section{Used in}{
 \code{\link{WoSSearchAndExport}}
}

