% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eolUrlScrapeR.R
\name{interSpecificDuplicates}
\alias{interSpecificDuplicates}
\title{Finds rows that are due to a name being contained within another name}
\usage{
interSpecificDuplicates(urls, names)
}
\arguments{
\item{names}{a tibble with all names associated with their internalTaxonIds}

\item{duplicates}{a dataframe of all the duplicated urls}
}
\value{
a vector of all the ids of species names that are contained within
other species names
}
\description{
Finds rows that are due to a name being contained within another name
}
