% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOCBirdList.R
\name{IOCAltnames}
\alias{IOCAltnames}
\title{Find species alternate names}
\usage{
IOCAltnames(formatted_IOC, cleaned_redlist_data)
}
\arguments{
\item{formatted_IOC}{tibble created by \code{\link{IOCDownload}} when
\code{tidy = TRUE}}

\item{cleaned_redlist_data}{cleaned redlist data output by
\code{\link{tidyRedlist::tidySpeciesData}}}
}
\value{
a tibble with columns \code{internalTaxonId}, \code{scientificName},
\code{name}, \code{language}, \code{main} and \code{source}. Effectivedly
\code{formatted_IOC} pivotted longer.
}
\description{
Find species alternate names
}
