% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOCBirdList.R
\name{IOCDownload}
\alias{IOCDownload}
\title{Downloads specified version of IOC bird list}
\usage{
IOCDownload(version, tidy)
}
\arguments{
\item{version}{the version of the list you wish to download, in format e.g.
version = '10.1'. If current version is desired, use version = 'current'}

\item{tidy}{logical, if TRUE, will run \code{\link{IOCformat}} on the downloaded
tibble and return a tidied copy}
}
\value{
a tibble containing an unprocessed version of the IOC bird list
downloaded
}
\description{
Downloads specified version of IOC bird list
}
