% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOCBirdList.R
\name{IOCFormat}
\alias{IOCFormat}
\title{Formats IOC bird list into tidy data}
\usage{
IOCFormat(IOC_sheet)
}
\arguments{
\item{IOC_sheet}{IOC bird list tibble, downloaded using
\code{\link{IOCDownload}}}
}
\value{
a tibble containing the tidies IOC bird list
}
\description{
Formats IOC bird list into tidy data
}
