% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOCBirdList.R
\name{IOCSpeciesNames}
\alias{IOCSpeciesNames}
\title{select target spp. from IOC birdlist and add IDs}
\usage{
IOCSpeciesNames(formatted_IOC, cleaned_redlist_data)
}
\arguments{
\item{formatted_IOC}{tibble created by \code{\link{IOCDownload}} when
\code{tidy = TRUE}}

\item{cleaned_redlist_data}{cleaned redlist data output by
\code{\link{tidyRedlist::tidySpeciesData}}}
}
\value{
formatted_IOC for species in cleaned_redlist_data with
internalTaxonId and scientificName columns
}
\description{
select target spp. from IOC birdlist and add IDs
}
