% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WoSSearch.R
\name{keysToSearchBox}
\alias{keysToSearchBox}
\title{Find an element and enter keys}
\usage{
keysToSearchBox(remDr, method, path, keys)
}
\arguments{
\item{remDr}{remote driver for rselenium}

\item{method}{the type of path to be used for element selection}

\item{path}{path for element to be clicked}

\item{keys}{the keys to be entered into the element}
}
\description{
finds element specified by \code{xpath} and enters \code{keys}
}
\section{Used in}{
 \code{\link{WoSExport}}
}

