% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikiPageViews.R
\name{lastDayOfMonth}
\alias{lastDayOfMonth}
\title{find end day}
\usage{
lastDayOfMonth(end_year, end_month)
}
\arguments{
\item{end_year}{the end year to be considered}

\item{end_month}{the end month to be considered}
}
\description{
finds last day of the month
}
