% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSearchTerms.R
\name{originalIDOrder}
\alias{originalIDOrder}
\title{Order by original ID order}
\usage{
originalIDOrder(table_to_order, names)
}
\arguments{
\item{table_to_order}{the table/tibble that needs to be reordered}

\item{names}{the original tibble of names for each thing}
}
\value{
the data from \code{table_to_order} reordered such that the IDs are in the
same order as in \code{names}
}
\description{
Order by original ID order
}
