% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlapCheck.R
\name{overlapCheck}
\alias{overlapCheck}
\title{checks names for overlap which might result in the search for one species
returning another}
\usage{
overlapCheck(names)
}
\arguments{
\item{names}{the tidyr::tibble of names etc.}
}
\value{
names with three extra columns \code{contained_in} a list
of any names that contain the name in the row, \code{contains} a list of all
the names the name in the row contains, and \code{recommendation} with keep
(names with no overlap), possibly unnecessary (contains another name for the)
same species, or warn (contained within the name of another species)
}
\description{
checks names for overlap which might result in the search for one species
returning another
}
