% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlapCheck.R
\name{overlappingNames}
\alias{overlappingNames}
\title{find names that contained at least one other name and return the number of
the name they contain}
\usage{
overlappingNames(names)
}
\arguments{
\item{names}{tidyr::tibble of all the names for each item}
}
\value{
logical vector where true means that name is contained within at
least one other name
}
\description{
find names that contained at least one other name and return the number of
the name they contain
}
