% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOCBirdList.R
\name{removeDuplicatedNames}
\alias{removeDuplicatedNames}
\title{Remove duplicated names}
\usage{
removeDuplicatedNames(names)
}
\arguments{
\item{names}{tibble or df with at least the columns \code{internalTaxonId},
\code{name} and \code{main}}
}
\value{
a tibble with columns \code{internalTaxonId}, \code{scientificName},
\code{name}, \code{main} and \code{source}. Effectivedly
\code{formatted_IOC} pivotted longer.
}
\description{
Removes duplicated names within a species irrespective of whether source
is different, will preferentially remove from sources where
\code{main} = FALSE
}
