% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikiScrapeR.R
\name{splitWikiScrape}
\alias{splitWikiScrape}
\title{split text scraped from wikipedia pages into the constituent sections}
\usage{
splitWikiScrape(text, contents)
}
\arguments{
\item{text}{vector of texts from wikipedia pages}

\item{contents}{vector of contents sections corresponding to texts}
}
\value{
a list with one element per page and, within that, one element
per section
}
\description{
split text scraped from wikipedia pages into the constituent sections
}
