% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikiScrapeR.R
\name{textListToTibble}
\alias{textListToTibble}
\title{Convert a list of text into a tibble}
\usage{
textListToTibble(split_texts, pictures, url, ID)
}
\arguments{
\item{list}{a list of texts where each element is the text for one
species and each sub-element is a section of that text}

\item{scientificNames}{the scientific names to be searched}

\item{internalTaxonId}{an identifier for each species}
}
\value{
tibble with each section of the text as a row
}
\description{
Convert a list of text into a tibble
}
