% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eolUrlScrapeR.R
\name{urlFromDropdown}
\alias{urlFromDropdown}
\title{Extracts text and url from each element in an EOL search dropdown}
\usage{
urlFromDropdown(remDr, best_match, i, latin_names, internalTaxonId)
}
\arguments{
\item{remDr}{rselenium remote driver}

\item{best_match}{logical, if true, will return the link for which the
text best matches the latin name searched, if false, will return all the
dropdown results for each search term}

\item{i}{the position in loop, set to 1 if not looping}
}
\value{
a tibble with columns \code{internalTaxonId}, \code{text} and
\code{link}
}
\description{
Extracts text and url from each element in an EOL search dropdown
}
\section{Requires}{
 \code{\link{rselenium}}
}

