% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikiPageViews.R
\name{wikiPageViews}
\alias{wikiPageViews}
\title{Wikipedi page views}
\usage{
wikiPageViews(
  wiki_urls,
  granularity = c("daily", "monthly"),
  access = "all-access",
  agent = "user",
  start_year,
  start_month = "01",
  start_day = "01",
  end_year,
  end_month = "12",
  end_day = NULL
)
}
\arguments{
\item{wiki_urls}{tibble with ID and the url of the wikipedia pages being
queried}

\item{granularity}{the temporal granularity you want JSON results to be
returned with must be one of 'daily' or 'monthly'}

\item{access}{the method of access to be counted, must be one of:
'all-access', desktop', 'mobile-app', 'mobile-web'}

\item{agent}{the type of access to be counted, must be one of:
'all-agents', 'user', 'spider', 'automated'}

\item{start_year}{the year you want to start from}

\item{start_month}{the month you want to start from, defaults to January
(01)}

\item{start_day}{the day you want to start from, defaults to the first
(01)}

\item{end_year}{the year you want to end with}

\item{end_month}{the month you want to end with, defaults to December
(12)}

\item{end_day}{the day you want to end with, defaults to the last
day of the month}
}
\description{
query the wikipedia api to find the number of page views
}
