% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WoSSearch.R
\name{WoSExport}
\alias{WoSExport}
\title{Navigate WoS export menu}
\usage{
WoSExport(remDr, hits)
}
\arguments{
\item{remDr}{remote driver for rselenium}

\item{hits}{the number of hits returned by the WoS search}
}
\value{
the number of files that have been exported
}
\description{
downloads all records returned by a search, with the parameters set in
\code{\link{WoSInitialExportSetup}} and \code{\link{WoSSubsequentExportSetup}}
}
\section{Used in}{
 \code{\link{WoSSearchAndExport}}
}

