% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WoSSearch.R
\name{WoSHits}
\alias{WoSHits}
\title{Carries out multiple searches on Web of Science and saves the number of
hits for each}
\usage{
WoSHits(search_terms, search_ids, directory, WoS_url)
}
\arguments{
\item{search_terms}{a vector/ column of WoS format search terms}

\item{search_ids}{a vector/ column the same length as \code{search_ids}
containing an ID to be used to identify search results to each search term}

\item{WoS_url}{the url for the WoS page with a search bar once you have logged on}
}
\description{
Carries out multiple searches on Web of Science and saves the number of
hits for each
}
