% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WoSSearch.R
\name{WoSInitialExportSetup}
\alias{WoSInitialExportSetup}
\title{Select parameters in WoS export dropdowns}
\usage{
WoSInitialExportSetup(remDr)
}
\arguments{
\item{remDr}{remote driver for rselenium}
}
\description{
selects full record and BibTeX as export parameters, suitable for the first
time this menu is navigates
}
\section{Used in}{
 \code{\link{WoSExport}}
}

