% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WoSSearch.R
\name{WoSSearch}
\alias{WoSSearch}
\title{Search Web of Science}
\usage{
WoSSearch(search_terms, search_ids, directory, WoS_url)
}
\arguments{
\item{search_terms}{a vector/ column of WoS format search terms}

\item{search_ids}{a vector/ column the same length as \code{search_ids}
containing an ID to be used to identify search results to each search term}

\item{directory}{the directory to be used for downloaded search results}
}
\value{
the number of hits for each search term, the directory specified in
\code{directory} populated with search results
}
\description{
Search Web of Science
}
