% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WoSSearch.R
\name{WoSSearchAndExport}
\alias{WoSSearchAndExport}
\title{Search and export from WoS}
\usage{
WoSSearchAndExport(remDr, search_terms, search_ids, WoS_url)
}
\arguments{
\item{remDr}{remote driver for rselenium}

\item{search_terms}{a vector/ column of WoS format search terms}

\item{search_ids}{a vector/ column the same length as \code{search_ids}
containing an ID to be used to identify search results to each search term}
}
\value{
a vector with the number of files exported for each search
}
\description{
Searches each of the search terms from \code{search_terms} in turn and
exports their results to chrome's download directory
}
\section{Used in}{
 \code{\link{WoSSearch}}
}

