% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WoSSearch.R
\name{WoSSubsequentExportSetup}
\alias{WoSSubsequentExportSetup}
\title{Select parameters in WoS export dropdowns}
\usage{
WoSSubsequentExportSetup(remDr)
}
\arguments{
\item{remDr}{remote driver for rselenium}
}
\description{
selects full record and BibTeX as export parameters, suitable for the second
or subsequent times this menu is navigated
}
\section{Used in}{
 \code{\link{WoSExport}}
}

