% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikiScrapeR.R
\name{splitWikiScrape}
\alias{splitWikiScrape}
\title{Find an element and get text}
\usage{
splitWikiScrape(text, contents)
}
\arguments{
\item{remDr}{the remote driver}

\item{using}{the type of path to use}

\item{path}{the path to use}
}
\value{
a value containing the text from the element
}
\description{
Find an element and get text
}
